+++ src/frontend/src/views/SearchView.vue
@@ -1,11 +1,11 @@
 <script setup lang="ts">
-import { computed, onMounted, ref } from 'vue'
+import { onMounted, ref } from 'vue'
 
 import Header from '../components/organisms/Header.vue'
 import SearchBar from '@/components/SearchBar.vue'
 import CardItem from '@/components/CardItem.vue'
+import Pagination from '@/components/Pagination.vue'
 import { useSearchStore } from '@/stores/searchStore'
-import { useDisplay } from 'vuetify'
 
 const searchStore = useSearchStore()
 
@@ -15,27 +15,8 @@ const handleScroll = () => {
   showScrollButton.value = window.scrollY > 300
 }
 
-// ページトップへスクロール
-const scrollToTop = () => {
-  window.scrollTo({ top: 0, behavior: 'smooth' })
-}
-
 onMounted(() => {
   window.addEventListener('scroll', handleScroll)
-})
-
-const loadMore = () => {
-  searchStore.nextPage()
-  scrollToTop()
-}
-
-const { xs, sm, md } = useDisplay()
-
-const paginationVisible = computed(() => {
-  if (xs.value) return 3
-  if (sm.value) return 5
-  if (md.value) return 7
-  return 10
 })
 </script>
 
@@ -72,19 +53,10 @@ const paginationVisible = computed(() => {
       </div>
 
       <!-- ページネーション -->
-      <div class="pagination-container">
-        <v-pagination
-          v-model="searchStore.currentPage"
-          class="my-1"
-          :length="searchStore.totalPage"
-          :total-visible="paginationVisible"
-          style="color: darkgreen"
-          @click="loadMore"
-        ></v-pagination>
-      </div>
+      <Pagination />
 
       <!-- トップへ移動ボタン -->
-      <v-btn v-if="showScrollButton" class="scroll-to-top" @click="scrollToTop" icon>
+      <v-btn v-if="showScrollButton" class="scroll-to-top" @click="searchStore.scrollToTop()" icon>
         <v-icon>mdi-arrow-up</v-icon>
       </v-btn>
     </v-main>
